requirements.txt
asgiref==3.7.2
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==3.12.3
pytz==2024.1
PyYAML==6.0.1
setuptools==69.1.1
sqlparse==0.2.4
tzdata==2024.1
uritemplate==4.1.1
